[MRESOLVER-440] Clean up transport names
---

https://issues.apache.org/jira/browse/MRESOLVER-440
