+++ maven-resolver-transport-apache/src/test/resources/ssl/README.txt
