+++ maven-resolver-transport-apache/src/test/resources/ssl/client-store
