+++ maven-resolver-transport-apache/src/test/resources/ssl/server-store
