+++ pom.xml
@@ -58,7 +58,7 @@
     <module>maven-resolver-transport-file</module>
     <module>maven-resolver-transport-jetty</module>
     <module>maven-resolver-transport-jdk</module>
-    <module>maven-resolver-transport-http</module>
+    <module>maven-resolver-transport-apache</module>
     <module>maven-resolver-transport-wagon</module>
     <module>maven-resolver-supplier</module>
     <module>maven-resolver-demos</module>
